requirements.txt
# MGA AI Agent - Dependencies
# Colombian MGA Document Generator

# ══════════════════════════════════════════════════════════════
# Core LLM Support (Multi-model)
# ══════════════════════════════════════════════════════════════
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-groq>=0.0.1
google-generativeai>=0.3.0

# ══════════════════════════════════════════════════════════════
# Web Interface
# ══════════════════════════════════════════════════════════════
streamlit>=1.30.0

# ══════════════════════════════════════════════════════════════
# Document Generation & Extraction
# ══════════════════════════════════════════════════════════════
python-docx>=1.1.0
docx2pdf>=0.1.8
pywin32>=306; sys_platform == 'win32'

# Extraction dependencies
pymupdf>=1.23.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0

# ══════════════════════════════════════════════════════════════
# Data Visualization (Optional - for PIB graphs)
# ══════════════════════════════════════════════════════════════
matplotlib>=3.7.0

# ══════════════════════════════════════════════════════════════
# Database
# ══════════════════════════════════════════════════════════════
sqlite-utils>=3.35.0

# ══════════════════════════════════════════════════════════════
# Environment & Config
# ══════════════════════════════════════════════════════════════
python-dotenv>=1.0.0

# ══════════════════════════════════════════════════════════════
# PDF Processing (for update mode)
# ══════════════════════════════════════════════════════════════
PyPDF2>=3.0.0

# ══════════════════════════════════════════════════════════════
# Testing
# ══════════════════════════════════════════════════════════════
pytest>=7.0.0
